% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_All_TC_AuroraExecution.R
\name{Create_All_TC_EjecucionDeAurora}
\alias{Create_All_TC_EjecucionDeAurora}
\title{Create All TC - Ejecucion de Aurora}
\usage{
Create_All_TC_EjecucionDeAurora()
}
\value{
The function creates and makes available the databases in .rds
format for the ThugChem package.
}
\description{
Create All TC - Ejecucion de Aurora
}
\examples{
if(FALSE) Create_All_TC_EjecucionDeAurora()
}
