% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DataTC_02_Elements.R
\name{Create_DataTC_02_Elements}
\alias{Create_DataTC_02_Elements}
\title{Create DataTC 02 - Elements}
\usage{
Create_DataTC_02_Elements()
}
\value{
It generates a data.frame where each row is an element of
the periodic table to which more columns have been added that
will be used for different utilities.
For example, the 'Status_oxyde' column will allow you to filter the
database to only have available elements that are susceptible to
forming oxides.
}
\description{
Create DataTC 02 - Elements
}
\examples{
if(FALSE) Create_DataTC_02_Elements()
}
