% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DataTC_03_Valences.R
\name{Create_DataTC_03_Valences}
\alias{Create_DataTC_03_Valences}
\title{Create DataTC 03 - Valences}
\usage{
Create_DataTC_03_Valences()
}
\value{
It generates a data.frame where each row is an element of the
periodic table with a particular valence in such a way that each element
appears in rows as many times as the number of valences it has.
Different columns have also been added.
}
\description{
Create DataTC 03 - Valences
}
\examples{
if(FALSE) Create_DataTC_03_Valences()

}
