% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DataTC_04_Oxyde.R
\name{Create_DataTC_04_Oxyde}
\alias{Create_DataTC_04_Oxyde}
\title{Create Data TC 04 - Oxyde}
\usage{
Create_DataTC_04_Oxyde()
}
\value{
It generates a data.frame where each row is an element of the
periodic table with a particular valence in such a way that each element
appears in rows as many times as the number of valences it has. We then have
a particular oxide in each row. Different columns have also been added,
such as "Chem_formula" that details the chemical formula of each oxide.
}
\description{
Create Data TC 04 - Oxyde
}
\examples{
if(FALSE) Create_DataTC_04_Oxyde()
}
