% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_PackTC_04_Oxyde.R
\name{Create_PackTC_04_Oxyde}
\alias{Create_PackTC_04_Oxyde}
\title{Create PackTC 04 - Oxyde}
\usage{
Create_PackTC_04_Oxyde()
}
\value{
Generates a list where each element is a list with everything
related to the resolution of a particular oxide. Each item in the list is
named with:
\itemize{
\item Order number of the oxide in DataTC_04_Oxyde, with 3 digits.
\item Element and Valencia
\item Chemical formula of oxide
Example of the nomenclature: 001_H1_H2O1.
}
}
\description{
Create PackTC 04 - Oxyde
}
\examples{
if(FALSE) Create_PackTC_04_Oxyde()
}
