% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTC_01_PeriodicTable_esp.R
\docType{data}
\name{DataTC_01_PeriodicTable_esp}
\alias{DataTC_01_PeriodicTable_esp}
\title{DataTC - Periodic Table - Spanish Version}
\format{
\subsection{\code{DataTC_01_PeriodicTable_esp}}{

A data frame with 118 rows and 24 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
DataTC_01_PeriodicTable_esp
}
\description{
It is a data.frame in which each row of the database is a chemical
element, and columns with information have been added to perform data
filtering.
}
\keyword{datasets}
