% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTC_02_Elements.R
\docType{data}
\name{DataTC_02_Elements}
\alias{DataTC_02_Elements}
\title{DataTC 02 - Elements}
\format{
\subsection{\code{DataTC_02_Elements}}{

A data frame with 118 rows and 24 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
DataTC_02_Elements
}
\description{
It is a data.frame in which each row of the database is a chemical
element, and columns with information have been added to perform data
filtering.
}
\keyword{datasets}
