% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTC_03_Valences.R
\docType{data}
\name{DataTC_03_Valences}
\alias{DataTC_03_Valences}
\title{DataTC 03 - Valences}
\format{
\subsection{\code{DataTC_03_Valences}}{

A data frame with 118 rows and 24 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
DataTC_03_Valences
}
\description{
It is a data.frame in which each row of the database is a chemical
element, and columns with information have been added to perform data
filtering.
}
\keyword{datasets}
