% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraDataTC_04_Oxyde.R
\docType{data}
\name{ExtraDataTC_04_Oxyde}
\alias{ExtraDataTC_04_Oxyde}
\title{ExtraDataTC 04 - Oxyde}
\format{
\subsection{\code{ExtraDataTC_04_Oxyde}}{

A data frame with 118 rows and 24 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
ExtraDataTC_04_Oxyde
}
\description{
Need a more description!.
}
\keyword{datasets}
