% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd2.R
\name{gcd2}
\alias{gcd2}
\title{Great Common Divisor (New Version)}
\usage{
gcd2(...)
}
\arguments{
\item{...}{Numbers}
}
\value{
Great Commun Divisor
}
\description{
This function take numbers and give you the great commond ivisor.
s a new version of the original gcd() funtion from 'FRACTION' packages.
The original gcd() is only for two numbers, and gdc2() can take any numbers.
}
\examples{
gcd2(2,4,8)
}
