% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DataTC_04_Oxyde.R
\name{Plot_DataTC_04_Oxyde}
\alias{Plot_DataTC_04_Oxyde}
\title{Plot Oxyde Resolution from Data ThugChem}
\usage{
Plot_DataTC_04_Oxyde(
  chem_symbol,
  element_valence,
  gas_status_element,
  verbose = TRUE
)
}
\arguments{
\item{chem_symbol}{The chemical symbol of the element}

\item{element_valence}{The valence of the element}

\item{gas_status_element}{If the element is a gas is TRUE else FALSE}

\item{verbose}{Logic value. Speaking about the processing}
}
\value{
A plot with all steps of resolution.
}
\description{
Plot Oxyde Resolution from Data ThugChem
}
\examples{
Plot_DataTC_04_Oxyde(chem_symbol = "H",
                     element_valence = 1,
                     gas_status_element = TRUE,
                      verbose = TRUE)

}
