% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PackTC_04_Oxyde.R
\name{Plot_PackTC_04_Oxyde}
\alias{Plot_PackTC_04_Oxyde}
\title{Plot Oxyde Resolution from Pack ThugChem}
\usage{
Plot_PackTC_04_Oxyde(chem_symbol, element_valence, verbose = TRUE)
}
\arguments{
\item{chem_symbol}{The chemical symbol of the element}

\item{element_valence}{The valence of the element}

\item{verbose}{Logic value. Speaking about the processing}
}
\value{
A plot with all steps of resolution.
}
\description{
Plot Oxyde Resolution from Pack ThugChem
}
\examples{
Plot_PackTC_04_Oxyde(chem_symbol = "Fe", element_valence = 2, verbose = T)
}
