% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_DataTC_04_Oxyde01.R,
%   R/Plot_PackTC_04_Oxyde01.R
\name{Plot_PackTC_04_Oxyde01}
\alias{Plot_PackTC_04_Oxyde01}
\title{Plot Oxyde Resolution from Pack ThugChem}
\usage{
Plot_PackTC_04_Oxyde01(chem_symbol, element_valence, verbose = TRUE, language)

Plot_PackTC_04_Oxyde01(chem_symbol, element_valence, verbose = TRUE, language)
}
\arguments{
\item{chem_symbol}{The chemical symbol of the element}

\item{element_valence}{The valence of the element}

\item{verbose}{Logic value. Speaking about the processing}
}
\value{
A plot with all steps of resolution.

A plot with all steps of resolution.
}
\description{
Plot Oxyde Resolution from Pack ThugChem

Plot Oxyde Resolution from Pack ThugChem
}
\examples{
Plot_PackTC_04_Oxyde01(chem_symbol = "Fe", element_valence = 2, verbose = T,
language = "esp",)
Plot_PackTC_04_Oxyde02(chem_symbol = "Fe", element_valence = 2, verbose = T,
                      language = "esp",)
Plot_PackTC_04_Oxyde01(chem_symbol = "Fe", element_valence = 2, verbose = T,
language = "esp",)
Plot_PackTC_04_Oxyde02(chem_symbol = "Fe", element_valence = 2, verbose = T,
                      language = "esp",)
}
