% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PackTC_04_Oxyde02.R
\name{Plot_PackTC_04_Oxyde02}
\alias{Plot_PackTC_04_Oxyde02}
\title{Plot Oxyde Resolution from Data ThugChem}
\usage{
Plot_PackTC_04_Oxyde02(chem_symbol, element_valence, verbose = TRUE, language)
}
\arguments{
\item{chem_symbol}{The chemical symbol of the element}

\item{element_valence}{The valence of the element}

\item{verbose}{Logic value. Speaking about the processing}

\item{gas_status_element}{If the element is a gas is TRUE else FALSE}
}
\value{
A plot with all steps of resolution.
}
\description{
Plot Oxyde Resolution from Data ThugChem
}
\examples{
Plot_DataTC_04_Oxyde(chem_symbol = "H",
                     element_valence = 1,
                     gas_status_element = TRUE,
                      verbose = TRUE)

}
