% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_oxyde_Classic.R, R/prefix_oxyde_Stock.R
\name{prefix_oxyde_Classic}
\alias{prefix_oxyde_Classic}
\title{Title}
\usage{
prefix_oxyde_Classic(language = "esp")

prefix_oxyde_Classic(language = "esp")
}
\arguments{
\item{language}{Select your language.}
}
\value{
A vector with all prefix for each valence of each elemento.

A vector with all prefix for each valence of each elemento.
}
\description{
Title

Title
}
\examples{
prefix_oxyde_Classic(language = "esp")
prefix_oxyde_Classic(language = "esp")
}
