% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Push_DataRaw_to_pkg_TC.R
\name{Push_DataRaw_to_pkg_TC}
\alias{Push_DataRaw_to_pkg_TC}
\title{Push Data Raw to packages ThugChem}
\usage{
Push_DataRaw_to_pkg_TC(data_raw_folder = "./data-raw/OriginalData/")
}
\arguments{
\item{data_raw_folder}{This is the folder from which the .csv files will be
taken to be converted to the .rds format and be part of the ThugChem package.}
}
\value{
No devuelve nada. Solo pasa los archivos .csv a .rds y los coloca en
la carpeta correcta.
}
\description{
Push Data Raw to packages ThugChem
}
\examples{
if(FALSE) Push_DataRaw_to_pkg_TC()
}
