% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resolution_Oxyde.R
\name{Resolution_Oxyde}
\alias{Resolution_Oxyde}
\title{Resolution for Oxyde}
\usage{
Resolution_Oxyde(chem_symbol, element_valence, gas_status_element)
}
\arguments{
\item{chem_symbol}{The chemical symbol of the element}

\item{element_valence}{The valence of the element}

\item{gas_status_element}{If the element is a gas is TRUE else FALSE}
}
\value{
The function
}
\description{
Resolution for Oxyde
}
\examples{
Resolution_Oxyde(chem_symbol = "H",
                 element_valence = 1,
                 gas_status_element = TRUE)
}
