% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ThugChemPage.R
\name{server_ThugChemPage}
\alias{server_ThugChemPage}
\title{ThugChemPage - Server}
\usage{
server_ThugChemPage(input, output, session)
}
\arguments{
\item{input}{cosas}

\item{output}{cosas}

\item{session}{mas cosas}
}
\value{
asdas
}
\description{
ThugChemPage - Server
}
\examples{
if(FALSE) server_ThugChemPage
}
