% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver_Oxyde.R
\name{Solver_Oxyde}
\alias{Solver_Oxyde}
\title{Solver for Oxyde}
\usage{
Solver_Oxyde(element_valence = 1, gas_status_element)
}
\arguments{
\item{element_valence}{The valence of the element}

\item{gas_status_element}{Logic value. If the elemnt is gas TRUE else FALSE.}
}
\value{
A matrix of the infile
}
\description{
Make the resolution for steciometry oxyde equation, step by step.
Each row is a step to obtein the final equation.
Return a data.frame, only with coefficients and subindex.
}
\examples{
Solver_Oxyde(element_valence = 1, gas_status_element = TRUE)
}
