% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Take_Nomenclature_Oxyde.R
\name{Take_Nomenclature_Oxyde}
\alias{Take_Nomenclature_Oxyde}
\title{Resolution for Oxyde}
\usage{
Take_Nomenclature_Oxyde(
  DataTC_04_Oxyde,
  chem_symbol,
  element_valence,
  gas_status_element,
  language
)
}
\arguments{
\item{chem_symbol}{The chemical symbol of the element}

\item{element_valence}{The valence of the element}

\item{gas_status_element}{If the element is a gas is TRUE else FALSE}
}
\value{
The function
}
\description{
Resolution for Oxyde
}
\examples{
Resolution_Oxyde(chem_symbol = "H",
                 element_valence = 1,
                 gas_status_element = TRUE)
}
